Split account-related Playwright E2E tests into a separate test set
