+++ package.json
@@ -10,10 +10,10 @@
     "test:integration": "cross-env SUITE=\"integration\" ./sh/test.sh",
     "test:install": "cross-env ./sh/install-tests.sh",
     "test:integration:stock": "cross-env SUITE=\"integration-stock\" ./sh/test.sh",
-    "test:ends": "cross-env SUITE=\"end-to-end\" ./sh/test.sh",
-    "test:e2e": "cross-env SUITE=\"end-to-end\" ./sh/test-e2e-pw.sh",
+    "test:e2e-all": "cross-env SUITE=\"end-to-end-all\" ./sh/test-e2e-pw-all.sh",
+    "test:e2e-account": "cross-env SUITE=\"end-to-end-account\" ./sh/test-e2e-pw-account.sh",
     "test:e2e-stock": "cross-env SUITE=\"end-to-end-stock\" ./sh/test-e2e-pw-stock.sh",
-    "test:e2e-trace": "cross-env SUITE=\"end-to-end\" ./sh/test-e2e-pw-trace.sh",
+    "test:e2e": "cross-env SUITE=\"end-to-end\" ./sh/test-e2e-pw.sh",
     "test:client-unit": "cross-env SUITE=\"client-unit\" ./sh/test.sh",
     "test:server-unit": "cross-env SUITE=\"server-unit\" ./sh/test.sh",
     "test:show-results": "./sh/test-show-results.sh",
