+++ sh/test-clean-reports.sh
@@ -5,4 +5,5 @@ rm -f ./test/integration/report
 rm -f ./test/server-unit/report
 rm -f ./test/integration-stock/report
 rm -f ./test/end-to-end-playwright/report
+rm -f ./test/end-to-end-playwright/report-account
 rm -f ./test/end-to-end-playwright/report-stock
