+++ test/end-to-end-playwright/account/accountStatement/account_statement.spec.js
@@ -1,6 +1,6 @@
 const { chromium } = require('@playwright/test');
 const { test, expect } = require('@playwright/test');
-const TU = require('../shared/TestUtils');
+const TU = require('../../shared/TestUtils');
 
 const AccountStatementCorePage = require('./account_statement.page');
